requirements.txt
fastapi
httpx
pydantic
python-dotenv
uvicorn[standard]